xwindow_set_state(): Use uint32_t instead of long
This function calls xcb_change_property() with format==32 which means that it is
called with a 32-bit integer. However, on 64-bit machines, sizeof(long) == 64
which means that this code did the wrong thing.

This wasn't noticed before because it only causes a real difference on
big-endian machines.

+++ xwindow.c
@@ -34,9 +34,9 @@
  * \param state The state to set.
  */
 void
-xwindow_set_state(xcb_window_t win, long state)
+xwindow_set_state(xcb_window_t win, uint32_t state)
 {
-    long data[] = { state, XCB_NONE };
+    uint32_t data[] = { state, XCB_NONE };
     xcb_change_property(globalconf.connection, XCB_PROP_MODE_REPLACE, win,
                         WM_STATE, WM_STATE, 32, 2, data);
 }
